Move sparse types into new header
